pyproject.toml
[project]
name = "deswoman"
version = "0.8.2"
description = "DESwoMAN is a software that detects precursors of de novo genes - newly expressed Open Reading Frames (neORFs) - based on transcriptome data and analyses their mutations within populations and/or species."
readme = "README.md"
authors = [
    { name = "Anna Grandchamp", email = "anna.grandchamp@inserm.fr" }
]
requires-python = ">=3.12"  # set to 3.12 for tkinter in debian-based docker image
dependencies = [
    "biopython>=1.85",
    "customtkinter>=5.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
deswoman = "deswoman.deswoman_runner:entry_point_main"

[dependency-groups]
lint = [
    "ruff>=0.9.2",
]